僵尸可以破坏器械了

+++ Board/Bullets/Bullet.cs
@@ -2,6 +2,6 @@ namespace MVE;
 
 public partial class Bullet : LawnEntity
 {
-    public float Damage { get; set; } = 20.0f;
+    public double Damage { get; set; } = 20.0f;
 
 }
